[buildfix] Build fixes for VC2022, dfgTest now builds and all tests pass
Compile error fixes:
     -Added one explicit conversion to dfgStr.utf16 to Windows-specific wchar_t tests
     -TypedCharPtrT needed operator== overload for comparing with type which differed only by Char_T constness.
     -Added cast to DFG_ASCII and DFG_UTF8 macros to take into account that in C++20 u8 creates char8_t array instead of char array.
    -In dfgTestNumeric.cpp converted one lambda into a function as with MSVC2022 C++20 mode it generated obscure syntax errors like "syntax error: '#pragma loop(no_vector)' was unexpected here; expected '}'" pointing to some
seemingly unrelated boost-header such as units\derived_dimension.hpp

Silenced warnings:
    -Added some casts to silence "warning C5055: operator '==': deprecated between enumerations and floating-point types"

+++ dfg/charts/operations.hpp
@@ -724,7 +724,7 @@ namespace operations
             if (argList.valueCount() != 3)
                 return privInvalidCreationArgsResult();
             const auto axis = axisStrToIndex(argList.value(0));
-            if (axis == axisIndex_invalid)
+            if (axis == static_cast<int>(axisIndex_invalid))
                 return privInvalidCreationArgsResult();
             ChartEntryOperation op(FilterOperation::operation);
 
@@ -932,7 +932,7 @@ namespace operations
             return ChartEntryOperation();
 
         const auto axis = axisStrToIndex(argList.value(0));
-        if (axis == axisIndex_invalid)
+        if (axis == static_cast<int>(axisIndex_invalid))
             return ChartEntryOperation();
 
         ChartEntryOperation op(&Formula::operation);
